pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mechanism-gwas"
version = "1.0.0"
description = "Mechanism-First Causal Graphs for Noncoding GWAS"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@institution.edu"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.10",
    "pyyaml>=6.0",
    "requests>=2.28",
    "tqdm>=4.65",
    "pyarrow>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.0.270",
    "mypy>=1.0",
]

[project.scripts]
mechanism-gwas = "src.cli:main"

[project.urls]
Homepage = "https://github.com/xxx/mechanism-gwas"
Documentation = "https://mechanism-gwas.readthedocs.io"
Repository = "https://github.com/xxx/mechanism-gwas"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "D", "UP"]
ignore = ["D100", "D104"]
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"
